pyproject.toml
[tool.poetry]
name = "la"
version = "0.1.0"
description = ""
authors = ["Roman T <sir.romka.t@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1.3"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
django-filter = "^22.1"
celery = "^5.2.7"
setuptools = "^65.6.3"
drf-yasg = "^1.21.4"
django-admin-multiple-choice-list-filter = "^0.1.1"
django-admin-rangefilter = "^0.9.0"
drf-nested-routers = "^0.93.4"
drf-writable-nested = "^0.7.0"
django-phonenumber-field = "^7.0.1"
django-extensions = "^3.2.1"
django-cors-headers = "^3.13.0"
phonenumbers = "^8.13.2"
ujson = "^5.6.0"
django-redis-sessions = "^0.6.2"
django-admin-inline-paginator = "^0.3.0"
django-tinymce = "^3.5.0"
django-cities-light = "^3.9.2"
random-password-generator = "^2.2.0"
django-celery-beat = "^2.4.0"
django-notifications-hq = "^1.7.0"
requests = "^2.28.2"
django-templated-email = "^3.0.1"
django-celery-results = "^2.5.0"
httpx = "^0.24.0"
adrf = "^0.1.1"
stripe = "^5.4.0"
pillow = "^9.5.0"
langcodes = {extras = ["data"], version = "^3.3.0"}
django-admin-sortable2 = "^2.1.8"
coreapi = "^2.3.3"
hashids = "^1.3.1"
python-dotenv = "^1.0.0"
pdfkit = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[tool.isort]
known_first_party = "apps, config, tests"
known_django = "django"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
